Drar ut logging til egen utility og rydder opp litt
