+++ Digipost.Signature.Api.Client.Core.Tests/nlog.config
@@ -9,14 +9,20 @@
     <!-- the targets to write to -->
     <targets>
         <!-- write logs to file -->
-        <target xsi:type="File" name="target1" fileName="c:\temp\console-example.log"
-                layout="${date}|${level:uppercase=true}|${message} ${exception}|${logger}|${all-event-properties}" />
-        <target xsi:type="Console" name="target2"
+        <target xsi:type="File"
+                name="fileTarget"
+                fileName="${specialfolder:folder=UserProfile}/logs/signature-api-client-dotnet/signature-api-client-dotnet.log"
+                layout="${date}|${level:uppercase=true}|${message} ${exception}|${logger}|${all-event-properties}"
+                archiveEvery="Day"
+                archiveNumbering="Date"
+                archiveDateFormat="yyyy-MM-dd"/>
+        <target xsi:type="Console"
+                name="consoleTarget"
                 layout="${date}|${level:uppercase=true}|${message} ${exception}|${logger}|${all-event-properties}" />
     </targets>
 
     <!-- rules to map from logger name to target -->
     <rules>
-        <logger name="*" minlevel="Trace" writeTo="target1,target2" />
+        <logger name="*" minlevel="Trace" writeTo="fileTarget,consoleTarget"/>
     </rules>
 </nlog>
