--- Digipost.Signature.Api.Client.Program/AppSettings.cs
@@ -1,9 +1,0 @@
-using System.Reflection.Metadata.Ecma335;
-
-namespace Digipost.Signature.Api.Client.Program
-{
-    public class AppSettings
-    {
-        public string Title { get; set; }
-    }
-}
