+++ Digipost.Signature.Api.Client.Program/Digipost.Signature.Api.Client.Program.csproj
@@ -3,20 +3,10 @@
     <ItemGroup>
         <PackageReference Include="Digipost.Signature.Api.Client.Direct" Version="5.0.2"/>
         <PackageReference Include="Digipost.Signature.Api.Client.Portal" Version="5.0.2"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.Logging" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.Logging.Console" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.1.1"/>
-        <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.1.1"/>
-        <PackageReference Include="NLog" Version="4.5.11"/>
-        <PackageReference Include="NLog.Extensions.Logging" Version="1.3.0"/>
-        <PackageReference Include="packtest.main" Version="1.0.0.1"/>
     </ItemGroup>
 
     <ItemGroup>
+        <ProjectReference Include="..\Digipost.Signature.Api.Client.Core.Tests\Digipost.Signature.Api.Client.Core.Tests.csproj"/>
         <ProjectReference Include="..\Digipost.Signature.Api.Client.Core\Digipost.Signature.Api.Client.Core.csproj"/>
     </ItemGroup>
